Exact test for tetraploid random mating

Conditions on sufficient statistics to calculate an exact p-value against the null of random mating.

tetexact(y, log_p = FALSE, frac = FALSE)

Arguments

y

The vector of genotype counts (of of length 5). y[k] is the number of individuals with genotype k-1.

log_p

A logical. Should we return the log p-value (TRUE) or not (FALSE)?

frac

A logical. Does y contain fractional genotypes (e.g. because y is calculated by summing up posterior probabilities)?

Details

If you have genotype likelihoods, I would first convert those to genotype posteriors (e.g. using ldsep::gl_to_gp()) to obtain gp, then run y <- colSums(gp, na.rm = TRUE). Then use the frac = TRUE?

Author

David Gerard and Karene Matoka Nana

Examples

set.seed(1)

# Generate data according to random mating
p <- c(0.5, 0.3, 0.2)
q <- stats::convolve(p, rev(p), type = "open")
y <- c(stats::rmultinom(n = 1, size = 200, prob = q))
tetexact(y = y)
#> [1] 0.5745608

# Generate data different from random mating
q <- c(0.4, 0.3, 0.2, 0.1, 0.0)
y <- c(stats::rmultinom(n = 1, size = 200, prob = q))
tetexact(y = y)
#> [1] 0.0108922

## small change when using fractional counts
y <- y + runif(5)
tetexact(y = y)
#> [1] 0.004522973